Prevent warnings on macOS
Changes to prevent unused variable/parameter and type comparison warnings.

+++ src/zxing/zxing/common/reedsolomon/ReedSolomonEncoder.cpp
@@ -17,7 +17,7 @@ ReedSolomonEncoder::ReedSolomonEncoder(Ref<GenericGF> field) :
 
 Ref<GenericGFPoly> ReedSolomonEncoder::buildGenerator(int degree)
 {
-    if (degree >= cachedGenerators_.size()) {
+    if ((unsigned long)degree >= cachedGenerators_.size()) {
         Ref<GenericGFPoly> lastGenerator = cachedGenerators_.at(cachedGenerators_.size() - 1);
         for (int d = cachedGenerators_.size(); d <= degree; d++)
         {
@@ -60,7 +60,7 @@ void ReedSolomonEncoder::encode(std::vector<byte> &toEncode, int ecBytes)
         toEncode[dataBytes + i] = 0;
     }
 
-	for (size_t i = 0; i < coefficients->size(); i++)
+    for (int i = 0; i < coefficients->size(); i++)
       toEncode[dataBytes + numZeroCoefficients + i] = coefficients[i];
 }
